Used FormManager to decouple some of the image downloading calls
